requirements.txt
fastapi[all]
SQLAlchemy
psycopg2-binary
alembic
bcrypt
python-jose[cryptography]
python-multipart
pydantic[email]
